Added create-react-app folder
I used "npx create-react-app folderName" in order to initialize React into the main folder, this command will also generate a couple of additional files which I will be deleting in the next commit. Also, create-react-app will generate a .gitignore file automatically to ignore the node_modules folder and a couple of other folders and files.